requirement.txt
Django==1.6.5
Pillow==2.6.1
South==1.0.1
dj-database-url==0.3.0
django-mptt==0.7.a
gunicorn==19.1.1
